feat(reports): integrate main parameterized report queries

+++ backend/src/main/java/com/pawnhop/backend/report/dto/ProblematicLoanReportDto.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 
 public record ProblematicLoanReportDto(
 
-        Long loanId,
+        Integer loanId,
         String lastName,
         String phone,
         BigDecimal amount,
